Add trailing path separator if it's a directory.
This fixes 'exclude_pfx' not excluding the directory itself but only the
files inside. (e.g. 'exclude_pfx: "/proc/"' would still report '/proc' as
changed)
It also makes the report more readable since now it's explicit whether a
path is a directory or a file.
This does not change the generated walk to keep it compatible with the
old ones.